Config validation tests (#3)
A test suite to validate `dritf.yaml` config file.
- lists contain no duplicate values
- lists are sorted alphabetically - it may look unnecessary but I
suppose it will help keeping an order when a list of supported services
grow.
- there are no unknown names
- region list is actual
- service and resource type lists are actual and match [CloudFormation
schemas](https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/resource-type-schemas.html)

Tests are in Kotlin :scream: It looks weird for a command-line DevOps
tool. That was just faster to develop - working with data collections in
Go is PITA, and downloading zip-files by Gradle is easier. It could be
rewritten in Go if required.

This Kotlin code is used during the development process only. The tool
is purely in Go.